+++ .gitignore
@@ -1,1 +1,1 @@
-./src/__pycache__
+src/__pycache__/
