+++ requirements.txt
@@ -1,0 +1,2 @@
+cv2==4.1.1
+pytesseract==0.3.4
