+++ result/im2.txt
@@ -1,0 +1,18 @@
+2 5 6 3 1 7 9 8 4 
+9 8 4 5 6 2 7 1 3 
+7 1 3 8 9 4 2 6 5 
+4 6 8 1 2 5 3 9 7 
+5 9 7 4 3 8 6 2 1 
+1 3 2 9 7 6 4 5 8 
+3 2 5 7 8 9 1 4 6 
+6 4 1 2 5 3 8 7 9 
+8 7 9 6 4 1 5 3 2 
+(1, 2)
+(2, 4)
+(3, 9)
+(4, 6)
+(5, 1)
+(6, 8)
+(7, 3)
+(8, 5)
+(9, 7)
