+++ result/tc1.txt
