+++ result/tc3.txt
